Refactor file names and directories
