+++ src/commands/list.ts
@@ -1,9 +1,9 @@
 import { resolveGithubToken } from "../config/env.js";
-import { resolveGitContext } from "../utils/gitContext.js";
-import { GithubService } from "../services/githubService.js";
-import { NoGitRepoError, InvalidRemoteUrlError } from "../utils/gitUtils.js";
-import { logError } from "../utils/logger.js";
-import {warn, bold, dim, success } from "../utils/color.js";
+import { resolveGitContext } from "../utils/git-context.utils.js";
+import { GithubService } from "../services/github.service.js";
+import { NoGitRepoError, InvalidRemoteUrlError } from "../utils/git.utils.js";
+import { logError } from "../utils/logger.utils.js";
+import {warn, bold, dim, success } from "../utils/color.utils.js";
 
 async function listCommand(): Promise<void> {
     try {
