+++ src/commands/update.ts
@@ -1,9 +1,9 @@
 import { resolveGithubToken } from "../config/env.js";
-import { resolveGitContext } from "../utils/gitContext.js";
-import { GithubService } from "../services/githubService.js";
-import { prompt } from "../utils/prompt.js";
-import { NoGitRepoError, InvalidRemoteUrlError } from "../utils/gitUtils.js";
-import { logError, logSuccess } from "../utils/logger.js";
+import { resolveGitContext } from "../utils/git-context.utils.js";
+import { GithubService } from "../services/github.service.js";
+import { prompt } from "../utils/prompt.utils.js";
+import { NoGitRepoError, InvalidRemoteUrlError } from "../utils/git.utils.js";
+import { logError, logSuccess } from "../utils/logger.utils.js";
 
 async function updateCommand(pullNumber: number): Promise<void> {
     try {
